refactor(github): extract shared repository support
